+++ src/com/maddyhome/idea/vim/group/copy/PutGroup.kt
@@ -51,6 +51,7 @@ import com.maddyhome.idea.vim.helper.fileSize
 import com.maddyhome.idea.vim.helper.moveToInlayAwareOffset
 import com.maddyhome.idea.vim.option.ClipboardOptionsData
 import com.maddyhome.idea.vim.option.OptionsManager
+import java.awt.datatransfer.DataFlavor
 import java.util.*
 import kotlin.math.abs
 import kotlin.math.max
@@ -328,11 +329,12 @@ class PutGroup {
     val origContent: TextBlockTransferable = setClipboardText(text.text, text.transferableData)
     val allContentsAfter = CopyPasteManager.getInstance().allContents
     val sizeAfterInsert = allContentsAfter.size
-    val firstItemAfter = allContentsAfter.firstOrNull()
     try {
       pasteProvider.performPaste(context)
     } finally {
-      if (sizeBeforeInsert != sizeAfterInsert || firstItemBefore != firstItemAfter) {
+      val textOnTop =
+        ((firstItemBefore as? TextBlockTransferable)?.getTransferData(DataFlavor.stringFlavor) as? String) != text.text
+      if (sizeBeforeInsert != sizeAfterInsert || textOnTop) {
         // Sometimes inserted text replaces existing one. E.g. on insert with + or * register
         (CopyPasteManager.getInstance() as? CopyPasteManagerEx)?.run { removeContent(origContent) }
       }